Budget professional mobile

+++ budget/budget.cpp
@@ -2,8 +2,8 @@
 #include <string>
 #include <string_view>
 
-#include "date.h"
 #include "budget_system.h"
+#include "date.h"
 
 #ifdef BUDGET_TEST
 #include "budget_test.h"
@@ -61,7 +61,7 @@ int main() {
 
     switch (ParseQueryType(query_type)) {
       case QueryType::COMPUTE_INCOME:
-        cout << static_cast<double>(budget_system.ComputeIncome(from, to)) << endl;
+        cout << budget_system.ComputeIncome(from, to) << endl;
         break;
       case QueryType::EARN:
         IncomeValue income_value;
